requirements.txt
streamlit
tensorflow-cpu
scikit-learn
xgboost
imbalanced-learn
pandas
numpy
plotly
fpdf2
shap
